Working example with adding one simple note.

+++ MidiCreator/SMF/Util/Note.h
@@ -10,13 +10,13 @@ namespace SMF
 	private:
 		NotePitch pitch;
 		uint8_t volume;
-		short duration;
+		int duration;
 
 	public:
-		Note(NotePitch pitch, uint8_t volume, short duration);
+		Note(NotePitch pitch, uint8_t volume, int duration);
 
 		inline const NotePitch& getPitch() { return this->pitch; }
 		inline const uint8_t& getVolume() { return this->volume; }
-		inline const short& getDuration() { return this->duration; }
+		inline const int& getDuration() { return this->duration; }
 	};
 }
